// src/app/interfaces/scatterplot.ts
export interface NeighborhoodData {
        name: string;
        population: number;
        density: number;
        singleparent_nb: number;
        singleparent_pct: number;
        avg_age: number;
        median_age: number;
        no_diploma_nb: number;
        no_diploma_pct: number;
        highschool_diploma_nb: number;
        highschool_diploma_pct: number;
        median_income: number;
        unemployment_rate: number;
        C90_C10_ratio: number;
        Gini_index: number;
        total_crime: number;
        crime_rate: number;
}